requirements.txt
# Backend
fastapi==0.101.1
uvicorn==0.23.2
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
pydantic==2.0.3
gdown==4.7.1
chardet==5.2.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
geopy==2.3.0
python-dotenv==1.0.0

# Data Processing
pyarrow==12.0.1

# Streamlit (para desenvolvimento/prototipagem)
streamlit==1.25.0
altair==5.0.1
plotly==5.15.0
streamlit-plotly-events==0.0.6
streamlit-folium==0.13.0
folium==0.14.0

# Maps & Visualizations
leafmap==0.23.3
contextily==1.3.0
geopandas==0.13.2
mapclassify==2.6.0